Split-sync support.
Send a dummy rate report so the server knows to expect them.

Rearrangements to message tracking to be a bit more consistent,
and so that we send rate reports with a rate of 0.0 for aircraft
that are sending ADS-B positions, but the positions aren't suitable
for synchronization. (This lets the server avoid trying to do mlat
on them)