Ignore messages with a magic mlat timestamp in Beast input.
MAGIC_MLAT_TIMESTAMP is 0xFF01A71A71A7ULL (it's got multiple
1A7s, get it?)

This should be unlikely to ever appear in real data (12MHz
timestamps will need to be running for 270 days to reach that,
and it's never a valid Radarcape GPS timestamp).

The intent here is that the mlat client will (eventually) inject
synthetic DF17 messages containing position results back to the
receiver e.g. dump1090. The receiver might loop those messages back
to the client. We must inhibit processing them or else (a) it'll
confuse the sequencing and (b) it would make the aircraft report
as ADS-B enabled, which it isn't.